QA-Chatbot/requirements.txt
langchain
python-dotenv
langchain_community
streamlit
langchain_core